Feat: Implement internal search functionality for employees and candidates, including search service and controller
