--- src/main/java/com/staffinity/recruiting/candidates/infrastructure/adapters/EmployeeClient.java
@@ -1,7 +1,0 @@
-package com.staffinity.recruiting.candidates.infrastructure.adapters;
-
-import java.util.List;
-
-public interface EmployeeClient {
-    List<Employee> getEmployees();
-}
