+++ src/main/java/com/staffinity/recruiting/internalSearch/application/usecases/InternalSearchUseCase.java
@@ -1,0 +1,11 @@
+package com.staffinity.recruiting.internalSearch.application.usecases;
+
+import com.staffinity.recruiting.internalSearch.application.dto.InternalSearchResponse;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface InternalSearchUseCase {
+    InternalSearchResponse searchInternalEmployees(UUID vacancyId);
+    List<InternalSearchResponse.MatchItemDto> searchInternalEmployees();
+}
