+++ src/main/java/com/staffinity/recruiting/internalSearch/infrastructure/adapters/employees/EmployeeClient.java
@@ -1,0 +1,7 @@
+package com.staffinity.recruiting.internalSearch.infrastructure.adapters.employees;
+
+import java.util.List;
+
+public interface EmployeeClient {
+    List<Employee> getEmployees();
+}
