--- src/main/java/com/staffinity/recruiting/recruiting/application/usecases/InternalSearchUseCase.java
@@ -1,9 +1,0 @@
-package com.staffinity.recruiting.recruiting.application.usecases;
-
-import com.staffinity.recruiting.recruiting.application.dto.InternalSearchResponse;
-
-import java.util.List;
-
-public interface InternalSearchUseCase {
-    List<InternalSearchResponse> searchInternalEmployees();
-}
