--- src/main/java/com/staffinity/recruiting/security/infrastructure/config/SecurityConfig.java
@@ -1,4 +1,0 @@
-package com.staffinity.recruiting.security.infrastructure.config;
-
-public class SecurityConfig {
-}
