+++ src/main/resources/db/migration/V1__create_vacancies_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE vacancies (
-    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    id UUID PRIMARY KEY,
     hiring_manager_id UUID,
     recruiter_id UUID,
     title VARCHAR(255) NOT NULL,
@@ -8,11 +8,11 @@ CREATE TABLE vacancies (
     location VARCHAR(255),
     remote_allowed BOOLEAN DEFAULT FALSE,
     seniority VARCHAR(50),
-    status VARCHAR(50) NOT NULL, -- Ej: OPEN, CLOSED, DRAFT
-    salary_min DECIMAL(10, 2),
-    salary_max DECIMAL(10, 2),
-    currency VARCHAR(3) DEFAULT 'USD',
+    status VARCHAR(50) NOT NULL,
+    salary_min DECIMAL(19, 2),
+    salary_max DECIMAL(19, 2),
+    currency VARCHAR(10),
     closed_at TIMESTAMP,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+    created_at TIMESTAMP NOT NULL,
+    updated_at TIMESTAMP NOT NULL
 );
