+++ src/main/resources/db/migration/V2__create_candidates_table.sql
@@ -1,18 +1,18 @@
 CREATE TABLE candidates (
-    id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+    id UUID PRIMARY KEY,
     vacancy_id UUID NOT NULL,
-    first_name VARCHAR(100) NOT NULL,
-    last_name VARCHAR(100) NOT NULL,
-    email VARCHAR(255) NOT NULL UNIQUE,
-    phone_number VARCHAR(50),
+    first_name VARCHAR(255) NOT NULL,
+    last_name VARCHAR(255) NOT NULL,
+    email VARCHAR(255) NOT NULL,
+    phone_number VARCHAR(255),
     resume_url VARCHAR(500),
     linkedin_url VARCHAR(500),
     portfolio_url VARCHAR(500),
     application_status VARCHAR(50) NOT NULL,
-    source VARCHAR(100),
+    source VARCHAR(255),
     rejection_reason TEXT,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    created_at TIMESTAMP NOT NULL,
+    updated_at TIMESTAMP NOT NULL,
     CONSTRAINT fk_vacancy
         FOREIGN KEY (vacancy_id)
             REFERENCES vacancies (id)
